+++ CHANGELOG.md
@@ -481,3 +481,4 @@ All notable changes to Project Genesis will be documented in this file.
 - Implemented `is_water_dancer` trait. Entities with this trait gain energy when on `water` terrain.
 
 - Implemented `is_forest_dancer` trait. Entities with this trait gain energy when on `forest` terrain.
+- Implemented `is_ash_dancer` trait. Entities with this trait gain energy when on `ash` terrain.